superblock: Overhaul for dynamic trait object
Despite not liking these, the size of the structs
we're dealing with forces dynamic allocation (boxed)
anyway.

So, simplify the APIs, enforce a consistent codepath
through the trait, and merge the error types into one.
While we're at it, some renames like `btrfs::from_reader`
to cleanup the code in general.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>