Remove existing topology solution
While it basically does do the job it certainly doesn't account for
the difference in types of block devices, such as whether it's a
real device partition, a raw device, volume group, etc.

In order to make life easier for ourselves we'll need to go a bit
further and add a nesting doll kind of BlockDevice to gain the full
composite view, and discovery device-mapper/dm-crypt.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>