blsforme: Add basic scaffolding for sync op
In order to make `sync` a useful operation we initially need to get
the whole "can haz working bootloader" part working first. So we now
have an API for bootloader abstraction, which is seeded by some set
of input files.

Next we'll add the case insensitive path_join APIs we so needed back
in the CBM days, and add helpers to atomically write-if-different any
files to ESP/XBOOTLDR.

At that point we can finish sync of bootloader, and then retrofit the
kernel syncs.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>