blsforme: Add PathExt trait to automatically join_insensitive paths
This is absolutely required for interaction with the ESP whereby some
other tool may have created paths such as "efi/BOOT/" and we ask for "EFI/Boot",
and get into the situation of the case sensitive path not existing, but then
hit issues when manipulating that file down the road.

Relying on the filesystem being mounted as case insensitive isn't feasible as
we don't always control the mountpoint, nor is it a recommended mount option.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>