blsforme/file_utils: Add file comparison utilities
When dealing with FAT we really want to avoid writing unless it
is absolutely necessary. To facilitate this caveat we utilise the
parallel/mmap variant of blake3 to compare a set of paths to
determine if they differ, and produce a reduced set of paths that
require updating.

Note, an absent source/dest is also the same as "needs update".

Signed-off-by: Ikey Doherty <ikey@serpentos.com>