bootloader: Some early scaffolding around kernel installs
This only grabs a partial loader config path right now but is effectively the bulk
of the code required to install kernels. We determine the installation base depending
on whether an XBOOTLDR is available, and generate suitable IDs for the config files.

Right now we're missing some information in the schema object, as its tied quite deeply
to the os-release file. A subsequent change will lock more of the schema to os-release
in order to generate unique entries on the ESP matching the existing CBM/Solus schema
while also permitting a more unique identifier for entries on blsforme schema for moss.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>