systemd_boot: Generate (in memory) configuration for loading
We don't yet write the file to disk as we need to do some comparison
logic between a *stream* and and a real file, so that'll take some
jiggering.

We also don't yet merge all system cmdline snippets as we don't
actually load them. However, that'll be up to the manager to load them
in appropriate order.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>