kernel: Allow kernel-specific cmdline snippets
These are vendor provided at source and shouldn't be disabled, as they're
intended for an exact kernel release. Right now the motivation is for out
of tree modules such as the nvidia proprietary driver, but this also allows
a kernel to ship its own cmdline file that cannot be overriden.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>